requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
certifi==2020.12.5
chardet==3.0.4
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
idna==2.10
IMAPClient==2.1.0
lxml==4.6.3
olefile==0.46
openai==0.6.3
pdfminer.six==20181108
Pillow==8.2.0
pycryptodome==3.10.1
PyPDF2==1.26.0
python-pptx==0.6.18
pytz==2021.1
requests==2.25.1
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
textract==1.6.3
tqdm==4.60.0
tzlocal==1.5.1
urllib3==1.26.4
xlrd==1.2.0
XlsxWriter==1.4.2
